쪼리퐁님 지적사항 수정
