+++ .gitignore
@@ -2,6 +2,7 @@ Client/node_modules
 Server/run.bat
 Server/node_modules
 Server/lib/node_modules
-Server/lib/Web/public/js
+Server/lib/Web/public/js/**
+!Server/lib/Web/public/js/jquery.js
 Server/lib/sub/auth.json
 Server/lib/sub/global.json
