+++ Server/lib/Game/master.js
@@ -26,6 +26,7 @@ var KKuTu = require('./kkutu');
 var GLOBAL = require("../sub/global.json");
 var Const = require("../const");
 var JLog = require('../sub/jjlog');
+var Secure = require('../sub/secure');
 
 var MainDB;
 
@@ -273,16 +274,7 @@ exports.init = function(_SID, CHAN){
 		
 		MainDB.users.update([ 'server', SID ]).set([ 'server', "" ]).on();
 		if(Const.IS_SECURED) {
-			const options = {};
-			if(Const.SSL_OPTIONS.isPFX == true) {
-				options.pfx = File.readFileSync(Const.SSL_OPTIONS.PFX);
-			} else {
-				options.key = File.readFileSync(Const.SSL_OPTIONS.PRIVKEY);
-				options.cert = File.readFileSync(Const.SSL_OPTIONS.CERT);
-				if(Const.SSL_OPTIONS.isCA == true) {
-					options.ca = File.readFileSync(Const.SSL_OPTIONS.CA);
-				}
-			}
+			const options = secure();
 			HTTPS_Server = https.createServer(options)
 				.listen(global.test ? (Const.TEST_PORT + 416) : process.env['KKUTU_PORT']);
 			Server = new WebSocket.Server({server: HTTPS_Server});
